# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NamThu Education API",
        "description": "Comprehensive API documentation for NamThu Education Laravel Backend",
        "contact": {
            "email": "support@namthuedu.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/teacher/blogs": {
            "get": {
                "tags": [
                    "Blog Management"
                ],
                "summary": "Get teacher blogs",
                "description": "Get list of blogs created by authenticated teacher",
                "operationId": "979d49eb391a96887c03b9acd01edd81",
                "responses": {
                    "200": {
                        "description": "Blogs retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Blog Management"
                ],
                "summary": "Create new blog",
                "description": "Create a new blog post (teacher only)",
                "operationId": "4b1a4fca3f0bd83fc5b22fc7e54f0d2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "blogName",
                                    "blogContent",
                                    "blogType",
                                    "blogCategory"
                                ],
                                "properties": {
                                    "blogName": {
                                        "type": "string",
                                        "example": "Grammar Tips for VSTEP"
                                    },
                                    "blogContent": {
                                        "type": "string",
                                        "example": "Detailed content about grammar..."
                                    },
                                    "blogType": {
                                        "type": "string",
                                        "enum": [
                                            "grammar",
                                            "tips",
                                            "vocabulary"
                                        ],
                                        "example": "grammar"
                                    },
                                    "blogCategory": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "blogUrl": {
                                        "type": "string",
                                        "example": "grammar-tips-vstep"
                                    },
                                    "blogThumbnail": {
                                        "type": "string",
                                        "example": "thumbnail.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blog created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teacher/blogs/{id}": {
            "get": {
                "tags": [
                    "Blog Management"
                ],
                "summary": "Get blog details",
                "description": "Get detailed information about a specific blog",
                "operationId": "0f35025a43d318ca032df0ddeeb59302",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog details retrieved successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Blog Management"
                ],
                "summary": "Update blog",
                "description": "Update blog information",
                "operationId": "b5e43b7e76acc72077f6fa13ac1c2685",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "blogName": {
                                        "type": "string"
                                    },
                                    "blogContent": {
                                        "type": "string"
                                    },
                                    "blogType": {
                                        "type": "string",
                                        "enum": [
                                            "grammar",
                                            "tips",
                                            "vocabulary"
                                        ]
                                    },
                                    "blogCategory": {
                                        "type": "integer"
                                    },
                                    "blogUrl": {
                                        "type": "string"
                                    },
                                    "blogThumbnail": {
                                        "type": "string"
                                    },
                                    "pStatus": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "draft"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog updated successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blog Management"
                ],
                "summary": "Delete blog",
                "description": "Delete a blog (soft delete)",
                "operationId": "fc71feb135db1681a33c066bd0b900f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog deleted successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}